+++ src/util/codec/mysql/mod.rs
@@ -60,7 +60,7 @@ mod duration;
 pub mod decimal;
 pub mod types;
 mod time;
-mod json;
+pub mod json;
 
 pub use self::duration::Duration;
 pub use self::decimal::{Decimal, Res, DecimalEncoder, DecimalDecoder, dec_encoded_len};
